Crate dec19x5

Represent a 64-bit decimal type taking 14 digits before the dot and 5 digits after the dot.

The calculation uses integer math to ensure accurate financial amounts.

Example

use dec19x5::Decimal;
use std::str::FromStr;

fn main() {
    // parse from string
    let d = Decimal::from_str("13.45331").unwrap();

    // write to string
    assert_eq!("13.45331", &format!("{}", d));

    // load an i32
    let d = Decimal::from(10i32);

    // multiple ops are supported.
    assert_eq!(d + Decimal::from(3i32), Decimal::from(13i32));
}

Structs

Decimal
A Decimal type for integer calculation of financial amount.

Enums

Error

Constants

MAX
Maximum value is 92 233 720 368 547.75807

MIN
Minimum value is -92 233 720 368 547.75808