Struct dec19x5::Decimal

pub struct Decimal(_);

A Decimal type for integer calculation of financial amount.

Note

The type is able to take 14 digits before the dot and 5 digits after the dot.

Implementations

impl Decimal

pub fn new_with_scale(value: i128, scale: u8) -> Self

Initialize a Decimal value using a value integer and a scale which represent the number of digit after the dot.

Panic

A scale greater than 18 will cause a panic.

pub fn is_zero(self) -> bool

Returns true if the decimal is zero.

pub fn round(self) -> Decimal

Returns the nearest integer to a number. Round half-way cases away from 0.0.

Example

use dec19x5::Decimal;

let f: Decimal = 3.3.into();
let g: Decimal = 3.6.into();
let h: Decimal = (-3.3).into();
let i: Decimal = (-3.6).into();

assert_eq!(f.round(), 3.into());
assert_eq!(g.round(), 4.into());
assert_eq!(h.round(), (-3).into());
assert_eq!(i.round(), (-4).into());

pub fn round_2(self) -> Decimal

round to 2 digits

Example

use dec19x5::Decimal;

assert_eq!("200.50000", &format!("{}", Decimal::from(200.49999).round_2()));
assert_eq!("200.48000", &format!("{}", Decimal::from(200.48499).round_2()));
assert_eq!("201.00000", &format!("{}", Decimal::from(200.99999).round_2()));

pub const fn scale(&self) -> u8

pub fn value(&self) -> i128

pub const fn zero() -> Self

Returns a Decimal value of zero.